requirements.txt

fastapi
uvicorn
httpx
azure-identity
msgraph-sdk
python-dotenv
pydantic
aiofiles
